REFACTOR: Form component restructure for readability
