+++ apps/meally.com.au/src/common/components/elements/recipe_elemnts/form_items/Ingredient.tsx
@@ -1,0 +1,13 @@
+const Ingredient = (props: any) => {
+  return (
+    <div className="flex flex-row items-center py-1 gap-1">
+      <input type="checkbox" />
+    </div>
+  );
+};
+
+const IngredientContainer = (props: any) => {
+  return <></>;
+};
+
+export { Ingredient, IngredientContainer };
