+++ apps/meally.com.au/src/pages/admin/recipe-form/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { TextField } from '@mui/material';
 import React, { useState } from 'react';
-import Form from '@components/elements/recipe_elemnts/Form';
+import Form from '@components/layouts/RecipeFromLayout';
 import RecipeService from '@lib/service/RecipeService';
 
 const recipe = {
